Extract sanWriter and common nvui functions into ui/nvui
