+++ ui/nvui/src/chess.ts
@@ -1,13 +1,14 @@
 import { h, VNode } from 'snabbdom';
-import { Pieces, Rank, File, files } from 'chessground/types';
+import { Dests, Pieces, Rank, File, files } from 'chessground/types';
 import { invRanks, allKeys } from 'chessground/util';
-//import { Api } from 'chessground/api';
+import { Api } from 'chessground/api';
 import { Setting, makeSetting } from './setting';
 import { parseFen } from 'chessops/fen';
 import { chessgroundDests } from 'chessops/compat';
 import { SquareName, RULES, Rules } from 'chessops/types';
 import { setupPosition } from 'chessops/variant';
 import { parseUci } from 'chessops/util';
+import { SanToUci, sanWriter } from './sanWriter';
 
 export type Style = 'uci' | 'san' | 'literate' | 'nato' | 'anna';
 export type PieceStyle = 'letter' | 'white uppercase letter' | 'name' | 'white uppercase name';
@@ -632,4 +633,43 @@ export function possibleMovesHandler(
     }
     return false;
   };
+}
+
+const promotionRegex = /^([a-h]x?)?[a-h](1|8)=\w$/;
+const uciPromotionRegex = /^([a-h][1-8])([a-h](1|8))[qrbn]$/;
+
+function destsToUcis(dests: Dests) {
+  const ucis: string[] = [];
+  for (const [orig, d] of dests) {
+    if (d)
+      d.forEach(function (dest) {
+        ucis.push(orig + dest);
+      });
+  }
+  return ucis;
+}
+
+function sanToUci(san: string, legalSans: SanToUci): Uci | undefined {
+  if (san in legalSans) return legalSans[san];
+  const lowered = san.toLowerCase();
+  for (const i in legalSans) if (i.toLowerCase() === lowered) return legalSans[i];
+  return;
+}
+
+export function inputToLegalUci(input: string, fen: string, chessground: Api): string | undefined {
+  const legalUcis = destsToUcis(chessground.state.movable.dests!),
+    legalSans = sanWriter(fen, legalUcis);
+  let uci = sanToUci(input, legalSans) || input,
+    promotion = '';
+
+  if (input.match(promotionRegex)) {
+    uci = sanToUci(input.slice(0, -2), legalSans) || input;
+    promotion = input.slice(-1).toLowerCase();
+  } else if (input.match(uciPromotionRegex)) {
+    uci = input.slice(0, -1);
+    promotion = input.slice(-1).toLowerCase();
+  }
+
+  if (legalUcis.includes(uci.toLowerCase())) return uci + promotion;
+  else return;
 }
